setup authorization for each route

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskStatus } from './task-status.enum';
 import { Task } from './task.entity';
 import { TaskRepository } from './task.repository';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,8 +15,11 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) {}
 
-    async getTaskById (id: number): Promise<Task>  {
-        const task = await this.taskRepository.findOne(id);
+    async getTaskById (
+        id: number,
+        user: User
+    ): Promise<Task>  {
+        const task = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
         if (!task) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -24,16 +28,23 @@ export class TasksService {
         return task;
     }
 
-    async getTasks(getTasksFilterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(getTasksFilterDto);
+    async getTasks(
+        getTasksFilterDto: GetTasksFilterDto,
+        user: User
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(getTasksFilterDto, user);
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await  this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User
+    ): Promise<Task> {
+        const task = await  this.getTaskById(id, user);
 
         task.status = status;
 
@@ -42,8 +53,11 @@ export class TasksService {
         return task;
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(
+        id: number,
+        user: User
+    ): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
